requirements.txt
requests
discord.py
mysql-connector-python